Move mouse_mode code inside qt subfolder
