+++ glue/viewers/profile/qt/mouse_mode.py
