+++ glue/viewers/profile/qt/tests/test_mouse_mode.py
